extract_to_gcs/requirements.txt
selenium
beautifulsoup4
google-cloud-storage
